feat: add copy subfolder

+++ src/fs/copy.js
@@ -7,20 +7,26 @@ const baseFolderPath = path.resolve(moduleDirname, 'files');
 const copyFolderPath = path.resolve(moduleDirname, 'files_copy');
 
 export const copy = async () => {
-    try {
-        const files = await fsProm.readdir(baseFolderPath);
-        await fsProm.mkdir(copyFolderPath);
 
-        for (const file of files) {
-            const baseFile = path.join(baseFolderPath, file);
-            const newFile = path.join(copyFolderPath, file);
-            await fsProm.copyFile(baseFile, newFile);
+    async function copyDir(baseFolderPath, copyFolderPath) {
+        try {
+            const files = await fsProm.readdir(baseFolderPath);
+            await fsProm.mkdir(copyFolderPath);
+
+            for await (const file of files) {
+                const baseFile = path.join(baseFolderPath, file);
+                const newFile = path.join(copyFolderPath, file);
+                if (await (await fsProm.stat(baseFile)).isFile()) {
+                    await fsProm.copyFile(baseFile, newFile);
+                } else {
+                    await copyDir(baseFile, newFile);
+                }
+            }
+        } catch (err) {
+            throw new Error('FS operation failed');
         }
-
-        console.log(`Folder is copied successfully.`);
-    } catch {
-        throw new Error('FS operation failed');
-    }
+    };
+    copyDir(baseFolderPath, copyFolderPath);
 };
 
 copy();
